Module phloem::blob

Provides the backend-agnostic data container.

This is the main module of Phloem and probably the reason, why you are using or interested in this library.

Overview

The main overview of the library gives a overview, to which I will just link here.

Opinions

As Phloems' Blob is hihgly opinionated for the use in a machine learning scenario, it has some restrictions and designs decisions. The following two are the most important.

diff:
Beside the fields data and shape, the Blob also has a field diff, which stores the gradients that the network might compute over the data stored in the data field.

Numeric:
As only numeric values can be processed by a Network, the type of the data that the Blob can store is restricted to the type Float, which is implemented for f32 and f64 Append reexported from the num crate.

Structs

Blob
Backend-agnostic data container.