Struct phloem::blob::Blob

pub struct Blob<T: Float> {
    // some fields omitted
}

Backend-agnostic data container.

Methods

impl<T> Blob<T> where T: Float

fn new() -> Blob<T>

Creates a new Blob in host Memory.

fn of_shape(device: Option<&DeviceType>, shape: &[usize]) -> Blob<T>

Creates a new Blob with specified shape on device.

If device is None, Blob will be created in host Memory.

fn reshape(&mut self, new_shape: &[usize])

Reshapes the Blob to the new shape given.

Results in recreation of the Blob if the capacity of the new shape doesn't match the one of the old shape. This destroys all data inside the blob

If the capacity of the old and new shape match, no data will be destroyed.

fn shape(&self) -> Vec<usize>

Returns the shape of the Blob.

fn shape_string(&self) -> String

Returns a String representation of the Blobs' shape.

The first numbers represent the size of the dimension. The last number in brackets defines the dimensionality of the Blob.

fn capacity(&self) -> usize

Returns the allocated capacity of the Blob.

fn data(&self) -> &SharedMemory<T>

Returns a reference to the data of the Blob.

fn mut_data(&mut self) -> &mut SharedMemory<T>

Returns a mutable reference to the data of the Blob.

fn diff(&self) -> &SharedMemory<T>

Returns a reference to the diff of the Blob.

fn mut_diff(&mut self) -> &mut SharedMemory<T>

Returns a mutable reference to the diff of the Blob.

fn apply_diff(&mut self)

Apply the diff to the data.

In machine learnig this is used when the blob represents weights in a network, and the comuted gradients should be applied to weights.

If no backend is provided, comutations are executed on the host.