Struct phloem::blob::Blob

pub struct Blob<T: Numeric> {
    // some fields omitted
}

Backend-agnostic data container.

Methods

impl<T> Blob<T> where T: Numeric

fn new() -> Blob<T>

Creates a new Blob

fn of_shape(new_shape: Vec<usize>) -> Blob<T>

Creates a new Blob with specified shape

fn reshape(&mut self, new_shape: Vec<usize>)

Reshapes the Blob to the new shape given

fn shape(&self) -> Vec<usize>

Returns the shape of the Blob

fn shape_string(&self) -> String

Returns a String representation of the Blobs' shape

The first numbers represent the size of the dimension. The last number in brackets defines the dimensionality of the Blob.

fn len(&self) -> usize

Returns the length of the Blob data

fn is_empty(&self) -> bool

Returns a boolean value whether the Blobs' data is empty or not

fn capacity(&self) -> usize

Returns the allocated capacity of the Blob data.

fn cpu_data(&self) -> &Vec<T>

Returns a pointer to the data of the Blob

fn mutable_cpu_data(&mut self) -> &mut Vec<T>

Returns a mutable pointer to the data of the Blob

fn cpu_diff(&self) -> &Vec<T>

Returns a pointer to the diff of the Blob

fn mutable_cpu_diff(&mut self) -> &mut Vec<T>

Returns a mutable pointer to the diff of the Blob